feat(projects/oss): move oss list to contents
