+++ src/contents/projects/oss/list.ts
@@ -1,26 +1,3 @@
-export type Genre = string;
-export type Project = {
-	name: string;
-	link?: string;
-	description?: string | null;
-	icon: string;
-};
-
-export type Projects = Record<Genre, Project[]>;
-
-export type GHRepo = {
-	repo: {
-		name: string;
-		description: string | null;
-		stars: number;
-		watchers: number;
-		forks: number;
-		createdAt: string;
-		pushedAt: string;
-		updatedAt: string;
-	};
-};
-
 // @unocss-include
 export default {
 	'Vim/Neovim': [
